% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT.R
\name{CIBERSORT}
\alias{CIBERSORT}
\title{Main functions}
\usage{
CIBERSORT(exprs, perm = 100, QN = TRUE)
}
\arguments{
\item{exprs}{heterogenous mixed expression}

\item{perm}{Number of permutations, default is 100}

\item{QN}{Perform quantile normalization or not (TRUE/FALSE)}
}
\description{
Main functions
}
