% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT.R
\name{CoreAlg}
\alias{CoreAlg}
\title{CIBERSORT R script v1.03
Note: Signature matrix construction is not currently available; use java version for full functionality.
Author: Aaron M. Newman, Stanford University (amnewman@stanford.edu)
License: http://cibersort.stanford.edu/CIBERSORT_License.txt
Core algorithm
X cell-specific gene expression
y mixed expression per sample}
\usage{
CoreAlg(X, y)
}
\description{
CIBERSORT R script v1.03
Note: Signature matrix construction is not currently available; use java version for full functionality.
Author: Aaron M. Newman, Stanford University (amnewman@stanford.edu)
License: http://cibersort.stanford.edu/CIBERSORT_License.txt
Core algorithm
X cell-specific gene expression
y mixed expression per sample
}
