% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT.R
\name{doPerm}
\alias{doPerm}
\title{do permutations
perm Number of permutations
X cell-specific gene expression
y mixed expression per sample}
\usage{
doPerm(perm, X, Y)
}
\description{
do permutations
perm Number of permutations
X cell-specific gene expression
y mixed expression per sample
}
