% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oliog2affy.R
\name{oligo2affy}
\alias{oligo2affy}
\title{Transform results of read.celfiles() in oligo package to AffyBatch format}
\usage{
oligo2affy(oligoRaw)
}
\arguments{
\item{oligoRaw}{results of read.celfiles() in oligo package}
}
\value{
an S4 object with AffyBatch format
}
\description{
Transform results of read.celfiles() in oligo package to AffyBatch format
}
\examples{
\donttest{
library(oligo)
cels <- list.files()
gse118370_raw <- read.celfiles(cels)
oliog2affy(gse118370_raw)
}
}
