% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_bg.correct.R
\name{pre_bg.correct}
\alias{pre_bg.correct}
\title{background correct using rma from affy package}
\usage{
pre_bg.correct(raw)
}
\arguments{
\item{raw}{raw}
}
\value{
background correct data
}
\description{
background correct using rma from affy package
}
