% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_impute.knn.R
\name{pre_impute.knn}
\alias{pre_impute.knn}
\title{impute missing value by knn}
\usage{
pre_impute.knn(raw)
}
\arguments{
\item{raw}{raw}
}
\value{
no missing value
}
\description{
impute missing value by knn
}
