% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_batch.effect.R
\name{qa_batch.effect}
\alias{qa_batch.effect}
\title{batch effect}
\usage{
qa_batch.effect(raw)
}
\arguments{
\item{raw}{raw data}
}
\value{
batch effect
}
\description{
batch effect
}
\examples{
\donttest{
library(oligo)
cels <- list.files()
gse118370_raw <- read.celfiles(cels)
qa_batch.effect(gse118370_raw)
}
}
