% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_raw.image.R
\name{qa_raw.image}
\alias{qa_raw.image}
\title{Draw gene raw data image}
\usage{
qa_raw.image(raw)
}
\arguments{
\item{raw}{raw data}
}
\value{
write image
}
\description{
Draw gene raw data image
}
\examples{
\donttest{
library(oligo)
cels <- list.files()
gse118370_raw <- read.celfiles(cels)
qa_raw.image(gse118370_raw)
}
}
