% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.GEO.series.matrix.R
\name{read.GEO.series.matrix}
\alias{read.GEO.series.matrix}
\title{read series matrix from GEO database}
\usage{
read.GEO.series.matrix(matrix)
}
\arguments{
\item{matrix}{series matrix}
}
\value{
dataframe
}
\description{
read series matrix from GEO database
}
\examples{
\donttest{
    read.GEO.series.matrix('GSE42872_series_matrix.txt')
}
}
